% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{bank_erode_fp}
\alias{bank_erode_fp}
\title{Estimate the cumulative bank erosion  during a flood hydrograph for a floodplain channel}
\usage{
bank_erode_fp(Q, t, n, D84, D50, W, S, H = 0, fun = "ec")
}
\arguments{
\item{Q}{a list of discharge values at regular time intervals defining the flood hydrograph (m3/s)}

\item{t}{time interval between discharge measurements (hrs)}

\item{n}{Manning's n value for the main channel}

\item{D84}{84th percentile of the surface grain size distribution (m)}

\item{D50}{50th percentile of the grain size distribution (m)}

\item{W}{water surface width at the beginning of time interval t}

\item{S}{energy gradient of the stream channel (m/m)}

\item{H}{effective rooting depth for vegetation}

\item{fun}{choice of transport function. Default is Eaton and Church ("ec")}

\item{Z}{height of the terrace above the channel bottom (m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function implements the GBEM model for estimating bank erosion using
the standard model for streams where the floodplain is assumed to be equal
to the critical depth based on the bed texture and slope. The function outputs a
data frame containing variables for time since the start of the hydrograph
(time), discharge values for each time (Q), the predicted bank erosion
distance for the time step (erosion), the volume of sediment transport that
the stream can convey during the timestep (vb), the critical depth for bank
erosion (dc), channel width at the end of each timestep (wi)
}
