% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{bank_erode_trc}
\alias{bank_erode_trc}
\title{Estimate the cumulative bank erosion during a flood hydrograph in an incized channel}
\usage{
bank_erode_trc(
  Q,
  t,
  n,
  D84,
  D50,
  W,
  S,
  H = 0,
  Z,
  fun = "ec",
  g = 9.81,
  rho_s = 2650,
  rho = 1000
)
}
\arguments{
\item{Q}{a list of discharge values at regular time intervals defining the flood hydrograph (m3/s)}

\item{t}{time interval between discharge measurements (hrs)}

\item{n}{Manning's n value for the main channel}

\item{D84}{84th percentile of the surface grain size distribution (m)}

\item{D50}{50th percentile of the grain size distribution (m)}

\item{W}{water surface width at the beginning of time interval t}

\item{S}{energy gradient of the stream channel (m/m)}

\item{H}{effective rooting depth for vegetation grassy banks, no trees / shrubs H = 0.35;  1 to 5percent tree / shrub cover H = 0.50; 5 to 50percent tree / shrub cover H =  0.90; more than 50percent tree / shrub cover H = 1.1}

\item{Z}{height of the terrace above the channel bottom (m)}

\item{fun}{choice of transport function. Default is Eaton and Church ("ec")}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
This function implements the GBEM model for estimating bank erosion using
a simple adjustment parameter to adapt the standard model for streams where
the channel is incised, and confined by a terrace. The function outputs a
data frame containing variables for time since the start of the hydrograph
(time), discharge values for each time (Q), the predicted bank erosion
distance for the time step (erosion), the volume of sediment transport that
the stream can convey during the timestep (vb), the critical depth for bank
erosion (dc), channel width at the end of each timestep (wi)
}
