% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{calc_depths}
\alias{calc_depths}
\title{Calculate flow depths}
\usage{
calc_depths(xs_data, wse, mkplt = F)
}
\arguments{
\item{xs_data}{a data frame containing $dist (distances) and $z (elevations)}

\item{wse}{elevation of the water surface of interest (m)}

\item{mkplt}{logical option to plot the data, default is FALSE}
}
\description{
this function uses a water surface elevation (WSE) and the location index for
the channel thalweg to calculate the depths in the main channel.  disconnected
areas outside the main channel that are below the WSE are located and assigned
a depth of NA. This is particularly important on a fan
}
