% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{compare_fun}
\alias{compare_fun}
\title{Compare transport functions}
\usage{
compare_fun(d, S, d84 = 2 * d, g = 9.81, rho = 1000, rho_s = 2650)
}
\arguments{
\item{d}{median bed sediment grain diameter (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d84}{84th percentile diameter of bed sediment (m). Defaults to 2 * d}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function compares the transport functions for a specified grain size
and channel gradient over a reasonable range of shear stress values using the
default values for all functions
}
