% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{crit_power}
\alias{crit_power}
\title{Estimating critical stream power for bed entrainment}
\usage{
crit_power(D50, d, use_ferguson = FALSE)
}
\arguments{
\item{D50}{median bed sediment grain diameter (m)}

\item{d}{flow depth (m)}
}
\description{
Using the equation published by Ferguson (2005) or Petit et al. (2005), this
function estimates the critical stream power as a function of grain size,
and flow depth. These empirical equations are valid for D50 ranging from
0.020 to 0.150 m
}
