% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{d_star}
\alias{d_star}
\title{Estimating  dimensionless particle size}
\usage{
d_star(d, Gs = 1.65, g = 9.81, kv = 1e-06)
}
\arguments{
\item{d}{median bed sediment grain diameter (m)}

\item{Gs}{the submerged specific gravity of sediment. Defaults to 1.65}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{kv}{kinematic viscosity of water. Defaults to 1e-6}
}
\description{
this function estimates the dimensionless particle size
}
