% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{eb}
\alias{eb}
\title{Estimating bedload transport using Einstein and brown (1950)}
\usage{
eb(R, S, d, g = 9.81, rho = 1000, rho_s = 2650, kv = 1e-06)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d}{median bed sediment grain diameter (m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{kv}{kinematic viscosity of water. Defaults to 1e-6}
}
\description{
This function uses a dimensionless form of Einstein and brown (1950) bedload
transport function for use with sand bed rivers and gravel bed rivers.
}
