% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{ec}
\alias{ec}
\title{Estimating bed material transport using Eaton and Church (2011)}
\usage{
ec(
  R,
  S,
  d,
  d84 = 2 * d,
  g = 9.81,
  rho = 1000,
  rho_s = 2650,
  U = u_ferg(R, S, d84, a1, a2, g),
  a1 = 6.5,
  a2 = 2.5,
  kv = 1e-06,
  t_crit = t_crit_soul(d, rho, rho_s, g, kv)
)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d}{median bed sediment grain diameter (m)}

\item{d84}{84th percentile diameter of bed sediment (m). Defaults to 2 * d}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{U}{mean flow velocity. Defaults to flows calculated using Ferguson (2007)}

\item{a1}{constant fit by Ferguson (2007). Defaults to 6.5}

\item{a2}{constant fit by Ferguson (2007). Defaults to 2.5}

\item{kv}{kinematic viscosity of water. Defaults to 1e-6}

\item{t_crit}{critical dimensionless shear stress, calculated using Soulsby (1997)}
}
\description{
This function uses a dimensionless stream power equation from Eaton and Church
(2011) to estimate bed material load transport in sand bed rivers and gravel
bed rivers.
}
