% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{est_mu}
\alias{est_mu}
\title{estimate relative bank strength}
\usage{
est_mu(H, d)
}
\arguments{
\item{H}{effective rooting depth for vegetation}

\item{d}{channel depths}
}
\description{
this is a utility function used by GBEM to determine the relative bank strength
given the channel depth and rooting depth. It returns a variable that is the
bank strength relative to the bed strengths.
}
