% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{estimate_capacity}
\alias{estimate_capacity}
\title{Estimate transport capacity for a hydrograph}
\usage{
estimate_capacity(xs_hyd, Q, t, tc = 0.08, mkplt = FALSE)
}
\arguments{
\item{xs_hyd}{a data frame created by estimate_hydraulics()}

\item{Q}{a vector containing the flows that define the hydrograph}

\item{t}{a vector that contains the corresponding times for each Q value (hours)}

\item{tc}{the threshold dimensionless shear stress for channel change}

\item{mkplt}{option to make a plot of the data, defaults to FALSE}
}
\description{
this function uses a data frame created by the estimate_hydraulics() function
to approximate the volume of bed material that could be transported during a
hydrograph. It also estimates the geomorphically active volume of transport
based on a threshold dimensionless shear stress.
}
