% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{estimate_hydraulics}
\alias{estimate_hydraulics}
\title{Estimate flow hydraulics for range of flood stages}
\usage{
estimate_hydraulics(
  xs_data,
  S,
  d,
  d84 = 2 * d,
  ymax = 3.5,
  dx = 0.05,
  t_crit = t_crit_soul(d),
  g = 9.81,
  rho = 1000,
  rho_s = 2650,
  a1 = 6.5,
  a2 = 2.5,
  kv = 1e-06,
  fun = "ec",
  rp = FALSE
)
}
\arguments{
\item{xs_data}{a data frame containing $dist (distances) and $z (elevations)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d}{median bed sediment grain diameter (m)}

\item{d84}{84th percentile diameter of bed sediment (m). Defaults to 2 * d}

\item{ymax}{the maximum flow depth to model, measured at the thalweg, defaults to 1.5 m}

\item{dx}{the increment by which to change stage. defaults to 0.05 m}

\item{t_crit}{critical dimensionless shear stress, calculated using Soulsby (1997)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{a1}{constant fit by Ferguson (2007). Defaults to 6.5}

\item{a2}{constant fit by Ferguson (2007). Defaults to 2.5}

\item{kv}{kinematic viscosity of water. Defaults to 1e-6}

\item{fun}{choice of bedload function: 'eb','ec', 'rk', 'mpm', 'vr','wp', 'yln'. Defaults to 'ec' ('mpm', 'wp', 'eb', 'yln', 'rk')}

\item{rp}{option for 'rk' select equation suited to riffle pool morphology, default is FALSE}
}
\description{
this function estimates hydraulic properties at regular WSE elevation intervals
these data can be used to interpolate values for target flow values.
}
