% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{find_d_crit}
\alias{find_d_crit}
\title{find the critical depth at which bank erosion will begin}
\usage{
find_d_crit(H, t_c84, S, g = 9.81, rho = 1000, tol = 1e-04)
}
\arguments{
\item{H}{effective rooting depth for vegetation}

\item{t_c84}{stress required to entrain the 84th percentile of the bed surface}

\item{S}{energy gradient of the stream channel (m/m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
this is a utility function used by GBEM to determine the typical floodplain
thickness above the channel bed.
}
