% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{find_q_b}
\alias{find_q_b}
\title{Calculating volumetric unit transport rate for a specified function}
\usage{
find_q_b(R, S, D50, D84, fun = "ec", g = 9.81, rho = 1000, rho_s = 2650)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{D50}{50th percentile of the grain size distribution (m)}

\item{D84}{84th percentile of the surface grain size distribution (m)}

\item{fun}{choice of transport function. Default is Eaton and Church ("ec")}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function takes the hydraulic data required to calculate the dimensionless
shear stress and uses a specified bedload transport function to calculate
the volumetric transport rate per unit width
}
