% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{flow_regime}
\alias{flow_regime}
\title{Display flow regime}
\usage{
flow_regime(station, window = 1)
}
\arguments{
\item{station}{a string containing the station name (e.g. '08GA022' or '14048000')}

\item{window}{(optional) the size of the window to use if you wish to smooth the flow regime by calculating a rolling average}
}
\description{
This function loads data from WSC or USGS stations and extracts the mean
daily flow for every day of the year, then generates a plot showing the
flow regime.
}
