% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{fractional_capacity}
\alias{fractional_capacity}
\title{Integrating  Wilcock and Crowe (2003) over a hydrograph}
\usage{
fractional_capacity(
  hydrograph,
  xs_data,
  S,
  gsd,
  Fsand,
  g = 9.81,
  rho = 1000,
  rho_s = 2650
)
}
\arguments{
\item{hydrograph}{a data frame created by scs_hydrograph() (Q in m3/s and t in hours)}

\item{xs_data}{a data frame containing topography expressed as $dist (distances) and $z (elevations)}

\item{S}{gradient of the water surface slope (m/m)}

\item{gsd}{a data frame describing the grain size distribution created with function wolman_gsd()}

\item{Fsand}{proportion of the surface that is covered by sand sized sediment (affects entrainment, typically varies from 0 to 0.2)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function integrates the Wilcock and Crowe transport function over a
hydrograph for a given cross section to estimate the total fractional
transport volumes for the flood event. Output includes the inputted GSD
information for each size class (Di, cpf, Dm, Fi), as well as the volumetric
transport rate (Qbi), and the fraction of the computed bedload in each size
class (Li)
}
