% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{gbem}
\alias{gbem}
\title{the GBEM function: gravel-bed river bank erosion model}
\usage{
gbem(
  Q,
  t,
  n,
  D84,
  D50,
  W,
  S,
  H = 0,
  fun = "ec",
  g = 9.81,
  rho_s = 2650,
  rho = 1000
)
}
\arguments{
\item{Q}{discharge carried by the stream (m3/s)}

\item{t}{time for which Q acts on the stream channel (hrs)}

\item{n}{Manning's n value for the main channel}

\item{D84}{84th percentile of the surface grain size distribution (m)}

\item{D50}{50th percentile of the grain size distribution (m)}

\item{W}{water surface width at the beginning of time interval t}

\item{S}{energy gradient of the stream channel (m/m)}

\item{H}{effective rooting depth for vegetation; grassy banks, no trees / shrubs H = 0.35; 1 to 5 percent tree / shrub cover H = 0.50; 5 to 50 percent tree / shrub cover H =  0.90; more than 50 percent tree / shrub cover H = 1.1}

\item{fun}{choice of transport function. Default is Eaton and Church ("ec")}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
This function implements the gravel bed bank erosion model (GBEM) that is the
core function used by STOCHASIM and the BGC stochastic bank erosion model.  It
limits the amount of bank erosion that can occur in a given timestep based on
the bedload transport capacity of the stream proximate to the channel banks.
The output from this function is a vector containing: (1) the predicted widening
based only on a stability criterion; (2) the predicted widening based on
the stability criterion, constrained by the bedload transport capacity; (4)
the volume of bedload that could be transported during the time step t in
the near-bank region; and (4) the critical depth for bank erosion
}
