% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{load_section}
\alias{load_section}
\title{Load cross section data}
\usage{
load_section(xs_nm, dx = NA, rev = FALSE, trim = FALSE, mkplt = FALSE)
}
\arguments{
\item{xs_nm}{path to a csv file with 3 columns containing the profile data}

\item{dx}{optional dist. increment to resample section at regular intervals (m)}

\item{rev}{option to reverse the direction in which distance is measured}

\item{trim}{option to identify left & right banks and trim section}

\item{mkplt}{option to make a plot of the data (original & resampled/trimed)}
}
\description{
This function loads profile data that is saved in a text file (XYZ format) and
calculates the distance along the profile from the first data point in the
file. The function assumes that the first row is header information, and that
the data in column 1 correspond to x coordinates (or eastings), column 2
corresponds to y coordinates (northings), and column 3 corresponds to elevation.
}
