% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{make_trapezoid}
\alias{make_trapezoid}
\title{Create a cross section based on measured trapezoid dimensions}
\usage{
make_trapezoid(w_top, w_bot, H_bank, Z_fp, dx = 0.1)
}
\arguments{
\item{w_top}{measured or estimated top width of the trapezoid (m)}

\item{w_bot}{measured or estimated bottom width of the trapezoid (m)}

\item{H_bank}{measures or estimated vertical height of the trapezoid (m)}

\item{dx}{a parameter controlling the spacing between exported data points on the section}

\item{z_fp}{elevation of the top of the trapezoid (i.e. the floodplain or terrace elevation, m)}
}
\description{
This function creates profile data including the distance along the profile
from the left bank.
}
