% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{max_min_timing}
\alias{max_min_timing}
\title{Display timing for annual daily maximum and minimum flows}
\usage{
max_min_timing(station, qual = 0.6)
}
\arguments{
\item{station}{a string containing the station name (e.g. '08GA022' or '14048000')}

\item{qual}{a value from 0 to 1 specifying how complete an annual dataset should be in order to estimate an annual maximum discharge value for that year}
}
\description{
This function loads data from WSC or USGS stations and extracts the annual
daily maximum flow and minimum flow, and then plots it against day of the year
so that peak flow and low flow timing can be quickly assessed
}
