% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{qb_conversion}
\alias{qb_conversion}
\title{Converting dimensionless flux to flux}
\usage{
qb_conversion(q_star, d, g = 9.81, Gs = 1.65)
}
\arguments{
\item{q_star}{dimensionless bedload flux}

\item{d}{median bed sediment grain diameter (m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{Gs}{the submerged specific gravity of sediment. Defaults to 1.65}
}
\description{
this function converts dimensionless bedload transport flux values to fluxes
in cubic meters per second per meter
}
