% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{quick_ffa_ann}
\alias{quick_ffa_ann}
\title{Estimate flood quantiles for annual daily maximum flow}
\usage{
quick_ffa_ann(station, rp = c(2, 5, 10, 20, 50, 100, 200), qual = 0.6)
}
\arguments{
\item{station}{a string containing the station name (e.g. '08GA022' or '14048000')}

\item{rp}{a vector containing the return periods for which to estimate flood quantiles}

\item{qual}{a value from 0 to 1 specifying how complete an annual dataset should be in order to estimate an annual maximum discharge value for that year}
}
\description{
This function loads data from WSC or USGS stations and fits a generalized
extreme value (GEV) distribution to the data using maximum likelihood
estimation.
}
