% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{rk}
\alias{rk}
\title{Estimating bed material transport using Recking (2013)}
\usage{
rk(R, S, d, d84 = 2 * d, g = 9.81, rho = 1000, rho_s = 2650, rp = FALSE)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d}{median bed sediment grain diameter (m)}

\item{d84}{84th percentile diameter of bed sediment (m). Defaults to 2 * d}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{rp}{option to select equation suited to riffle pool morphology, default is FALSE}
}
\description{
This function uses a dimensionless shear stress equation from Recking
based on the D84 of the bed surface, and is the version for use with
field data.
}
