% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{scs_hydrograph}
\alias{scs_hydrograph}
\title{Creating a hydrograph based on the Standard SCS Dimensionless Unit Hydrograph}
\usage{
scs_hydrograph(Qpeak, tpeak, dt)
}
\arguments{
\item{Qpeak}{the peak instantaneous discharge to be reached during the hydrograph (cumecs)}

\item{tpeak}{the lag time between the peak rainfall input and the peak runoff (hours)}

\item{dt}{a parameter that controls the time increment for data output (in hours)}
}
\description{
This function uses the Soil Conservation Service Standard Unit Hydrograph to
simulate a simple hydrograph associated with a rainfall event. The hydrograph
depends on a parameter controlling the peak flow that will be reached and a
parameter controlling how quickly the hydrograph will rise and fall. The
shape of the hydrograph remains the same.
}
\details{
The dimensionless unit hydrograph used by the SCS was developed by Victor
Mockus and was derived based on a large number of unit hydrographs from basins
that varied in characteristics such as size and geographic location. The unit
hydrographs were averaged and the final product was made dimensionless by
considering the ratios of Q/Qpeak (flow/peak flow) on the ordinate axis and
t/tpeak (time/time to peak) on the abscissa. This final, dimensionless unit
hydrograph, which is the result of averaging a large number of individual
dimensionless unit hydrographs, has a time-to-peak located at approximately
20 percent of its time base and an inflection point at 1.7 times the time-to-peak.
}
