% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{step_length}
\alias{step_length}
\title{Calculating the step length for sediment transport of coarse bed material}
\usage{
step_length(
  D50,
  q,
  d,
  S,
  dt,
  flashy = FALSE,
  use_ferguson = TRUE,
  g = 9.81,
  rho = 1000
)
}
\arguments{
\item{D50}{median bed sediment grain diameter (m)}

\item{q}{a list of specific discharge values for a hydrograph (m3/s/m)}

\item{d}{flow depth (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{dt}{time interval between discharge measurements (hrs)}

\item{flashy}{a logical variable (true if the hydrograph is short and flashy)}

\item{use_ferguson}{logical variable for selecting the ferguson equation for calculating the critical stream power}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
this function calculates the average distance for bedload transport during
a flow of a given duration
}
