% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{t_crit_soul}
\alias{t_crit_soul}
\title{Estimating critical dimensionless shear stress}
\usage{
t_crit_soul(d, rho = 1000, rho_s = 2650, g = 9.81, kv = 1e-06)
}
\arguments{
\item{d}{median bed sediment grain diameter (m)}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{kv}{kinematic viscosity of water. Defaults to 1e-6}
}
\description{
Using the equation published by Soulsby (1997), this function estimates the
critical dimensionless shear stress as a function of grain size.
}
