% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{t_ref_star}
\alias{t_ref_star}
\title{Estimating reference dimensionless shear stress for Wilcock & Crowe (2003)}
\usage{
t_ref_star(Fsand)
}
\arguments{
\item{Fsand}{Proportion of sand on the bed surface}
}
\description{
Using the equation published by Wilcock & Crow, this function estimates the
reference dimensionless shear stress based on the fraction of sand on the bed.
}
