% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{t_star}
\alias{t_star}
\title{Estimating the dimensionless shear stress}
\usage{
t_star(tau, d, g = 9.81, rho = 1000, rho_s = 2650)
}
\arguments{
\item{tau}{the cross section-averaged shear stress (Pa/m2).}

\item{d}{median bed sediment grain diameter (m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function calculates the dimensionless shear stress using the bed surface
median grain size.
}
