% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{tau_1D}
\alias{tau_1D}
\title{Calculating the cross section average shear stress}
\usage{
tau_1D(R, S, g = 9.81, rho = 1000)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
this function calculates the average shear stress acting on the channel
boundary for a given cross section.
}
