% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{u_ferg}
\alias{u_ferg}
\title{Estimating the mean flow velocity}
\usage{
u_ferg(R, S, d84, a1 = 6.5, a2 = 2.5, g = 9.81)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d84}{84th percentile diameter of bed sediment (m)}

\item{a1}{constant fit by Ferguson (2007). Defaults to 6.5}

\item{a2}{constant fit by Ferguson (2007). Defaults to 2.5}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}
}
\description{
This function uses the Ferguson (2007) flow resistance law to estimate
the mean velocity of the flow using the bed surface D84
}
