% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{unit_power}
\alias{unit_power}
\title{Calculating the cross section average unit stream power}
\usage{
unit_power(q, S, g = 9.81, rho = 1000)
}
\arguments{
\item{q}{specific discharge (m3/s/m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}
}
\description{
this function calculates the unit stream power acting on the channel
boundary for a given cross section.
}
