% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{wc}
\alias{wc}
\title{Estimating fractional bedload transport using Wilcock and Crowe (2003) in m3/second}
\usage{
wc(R, S, gsd, Fsand, g = 9.81, rho = 1000, rho_s = 2650)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{gsd}{a data frame describing the grain size distribution created with function wolman_gsd()}

\item{Fsand}{proportion of the surface that is covered by sand sized sediment (affects entrainment, typically varies from 0 to 0.2)}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}

\item{counts}{number of stones in each size class size class bounds (!!!must start with a 0!!!)}
}
\description{
This function uses the Wilcock and Crowe transport function on standard data
from a surface Wolman count to estimate the fraction-specific transport rate
for a given shear stress. Output is a data frame containing inputted GSD
information for each size class (Di, cpf, Dm, Fi), as well as the dimensionless
transport rate (Wi), the volumetric transport rate per unit width (qbi), and
the fraction of the computed bedload in each size class (Li)
}
