% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{wolman_gsd}
\alias{wolman_gsd}
\title{Processing Wolman count data into grain size distribution}
\usage{
wolman_gsd(Di, counts)
}
\arguments{
\item{Di}{a list of grain sizes representing the upper bounds of the size classes}

\item{counts}{the number of stones sampled in the szie class defined by the upper bound values, Di}
}
\description{
This function uses standard data from a surface Wolman count to estimate
cumulative percent finer for each size, the fraction of the distribution in
a given size class, and the median size of each size class. Output is a data
frame containing the upper limit of each size class (Di), the cumulative
proportion finer than the size listed in Di (cpf), the median size for each
size class (Dm), and the fraction of the distribution in each class (Fi)
}
