% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{sim_gsd}
\alias{sim_gsd}
\title{Processing Wolman count data into grain size distribution}
\usage{
sim_gsd(D50, sp = 1, Map = FALSE)
}
\arguments{
\item{D50}{an estimate of the median bed surface grain size (mm)}

\item{sp}{(optional) the standard deviation of the log-normal distribution (in phi units), default of 1 gives a D84 twice the D50}

\item{Map}{(optional) LOGICAL, when TRUE, the distribution is graphed}
}
\description{
This function uses standard data from a surface Wolman count to estimate
cumulative percent finer for each size, the fraction of the distribution in
a given size class, and the median size of each size class. Output is a data
frame containing the upper limit of each size class (Di), the cumulative
proportion finer than the size listed in Di (cpf), the median size for each
size class (Dm), and the fraction of the distribution in each class (Fi)
}
