% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedload_functions.R
\name{wp}
\alias{wp}
\title{Estimating bedload transport using Wong and Parker (2006)}
\usage{
wp(R, S, d, t_crit = 0.047, g = 9.81, rho = 1000, rho_s = 2650)
}
\arguments{
\item{R}{hydraulic radius for the flow (m)}

\item{S}{gradient of the water surface slope (m/m)}

\item{d}{median bed sediment grain diameter (m)}

\item{t_crit}{critical dimensionless shear stress. Defaults to 0.047}

\item{g}{the acceleration of gravity. defaults to 9.81 m/s2}

\item{rho}{fluid density. Defaults to 1000 kg/m3}

\item{rho_s}{sediment density. Defaults to 2650 kg/m3}
}
\description{
This function uses a dimensionless form of a modified version of Meyer Peter
and Muller equation to estimate the dimensionless bedload flux for a given
shear stress and characteristic grain size. Using the original Meyer Peter and Muller equation will produce higher
transport estimates and is more conservative than using this modified version.
}
